family_tree_project/requirements.txt
asgiref>=3.4.1
beautifulsoup4>=4.10.0
Django>=3.2.7
django-annoying>=0.10.6
django-bootstrap-v5>=1.0.5
django-braces>=1.14.0
django-cors-headers>=3.11.0
djangorestframework>=3.13.1
djangorestframework-recursive>=0.1.2
djangorestframework-simplejwt>=5.2.0
pyaml>=21.10.1
PyJWT>=2.4.0
pytz>=2021.1
PyYAML>=6.0
six>=1.16.0
soupsieve>=2.2.1
sqlparse>=0.4.2
